+++ changelog.md
@@ -5,6 +5,7 @@
 - Fixed the appearance of holograms in other worlds
 - Added command ``/fly get <player>``
 - Added command ``/fly info``
+- Added command ``/ess delete-world <world>``, allows you to delete data related to a world
 
 # 1.0.1.2